+++ src/main/java/bighouse/restapi/repository/RepositoryProduct.java
@@ -4,6 +4,7 @@ import bighouse.restapi.model.ModelProduct;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface RepositoryProduct extends CrudRepository<ModelProduct, Integer> {
+public interface RepositoryProduct extends CrudRepository<ModelProduct, Long> {
 }
